backend/requirements.txt
# ----------------------------------------------------
# Conteúdo para o arquivo: backend/requirements.txt
# ----------------------------------------------------

# Framework da API e Servidor
fastapi
uvicorn[standard]
pydantic

# Framework do Agente e Memória
langgraph[redis]
langchain
langchain_openai
langchain_community
openai
python-dotenv
redis

# Ferramentas de Dados e Busca
duckdb
pandas
google-search-results

# Ferramentas de RAG (Ingestão e Armazenamento)
pypdf
chromadb
tiktoken
pdf2image
Pillow
